modification Job6 jour 2
changement de la variant de int() en float + modification faute de frappe sur d'autre job

+++ jour02/job06.py
@@ -8,4 +8,4 @@ def nombre(nombre):
     else :
         print ("je suis le nombre 0")
 
-nombre(int(num))
+nombre(float(num))
